Allow fast finishes in Travis.
This means that if all builds that are not allowed_failures (pypy) are
successes, the build will immediately succeed, and that if any
noon-allowed_failure build fails, the entire build will immediately
fail.

This should make the tests be more responsive, and I expect that it will
still run all the tests, just after the failure notification goes out.